reoganized datatypes to group by topic
